apps/api/requirements.txt
# FastAPI Core
fastapi==0.116.2
uvicorn[standard]==0.32.0
gunicorn==23.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
sqlalchemy[asyncio]==2.0.36

# Redis
redis[hiredis]==5.2.0
aioredis==2.0.1

# Authentication & Security
pyjwt[crypto]==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Firebase Admin
firebase-admin==6.5.0

# AWS Services
boto3==1.35.16
aiobotocore==2.15.1

# HTTP Client
httpx==0.27.2
aiohttp==3.10.11

# Validation & Serialization
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# Email
python-jose[cryptography]==3.3.0

# Monitoring & Logging
sentry-sdk[fastapi]==2.17.0
structlog==24.4.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
pytest-mock==3.14.0

# Code Quality
ruff==0.7.4
mypy==1.13.0
black==24.10.0

# Environment Management
python-dotenv==1.0.1

# Image Processing
pillow==10.4.0

# Utilities
python-slugify==8.0.4
pytz==2024.2